pyproject.toml
[project]
name = "pinocchio-models"
version = "0.1.0"
description = "Pinocchio multibody models for classical barbell exercises with Gepetto, Pink IK, and Crocoddyl optimal control"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
keywords = ["pinocchio", "biomechanics", "simulation", "barbell", "strength-training", "urdf"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Physics",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]

dependencies = [
    "numpy>=1.26.4,<3.0.0",
    "scipy>=1.13.1",
    "lxml>=5.0.0",
]

[project.optional-dependencies]
pinocchio = [
    "pin>=3.0.0",
]
gepetto = [
    "gepetto-viewer-corba>=5.0.0",
]
pink = [
    "pin-pink>=3.0.0",
]
crocoddyl = [
    "crocoddyl>=2.0.0",
]
all-addons = [
    "pinocchio-models[pinocchio,gepetto,pink,crocoddyl]",
]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.0.0",
    "pytest-timeout>=2.0.0",
    "pytest-xdist>=3.0.0",
    "ruff>=0.14.0",
    "mypy>=1.15.0",
    "hypothesis>=6.0.0",
]

[project.urls]
Homepage = "https://github.com/D-sorganization/Pinocchio_Models"
Repository = "https://github.com/D-sorganization/Pinocchio_Models"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
include = ["src/**/*"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = [
    "E", "F", "I", "UP", "B", "T201", "SIM",
    "C4", "PIE", "PLE", "FURB", "RSE", "LOG",
    "PERF", "RET",
]
ignore = [
    "E501",
    "RET504",
    "PERF203",
    "PERF401",
]

[tool.ruff.lint.per-file-ignores]
"scripts/**" = ["T201"]
"tests/**" = ["T201"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
check_untyped_defs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = [".", "src"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-ra -q --strict-markers --strict-config --durations=20 -n auto --dist loadscope"
markers = [
    "slow: marks tests as slow",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "requires_pinocchio: tests requiring pinocchio package",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.coverage.run]
source = ["src"]
branch = true
omit = ["*/tests/*", "*/test_*", "*/__pycache__/*"]

[tool.coverage.report]
fail_under = 80
show_missing = true
skip_empty = true
precision = 1
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.",
    "raise NotImplementedError",
    "pass",
    "except ImportError",
    "if TYPE_CHECKING:",
]
